Clean the code into module
